improve build for both javac and normal build artifacts

+++ test-run.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+if [ -z $1 ]; then
+    echo "Specify either --checkout or --only-checkout"
+    exit 1
+fi
+
+suffix="$2"
+
 if [[ "$1" == "--checkout" || "$1" == "--only-checkout" ]]; then
     if [ -d ./extensions ]; then
         rm -Rf ./extensions
@@ -9,7 +16,7 @@ if [[ "$1" == "--checkout" || "$1" == "--only-checkout" ]]; then
     echo "## Checkout Phase"
     ./scripts/checkout-modules.sh
     echo "## Patch Phase"
-    ./scripts/merge-prs.sh true javac
+    ./scripts/merge-prs.sh true $suffix
 fi
 
 if [[ "$1" != "--only-checkout" ]]; then
